break up streaming execution context into smaller for testable interfaces and files.
